requirements.txt
google-adk
python-dotenv
requests
mcp
google-cloud-aiplatform
vertexai
llama-index
base58